+++ src/main/java/ru/practicum/shareit/booking/dto/BookingDto.java
@@ -1,12 +1,9 @@
 package ru.practicum.shareit.booking.dto;
 
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 import ru.practicum.shareit.booking.BookingStatus;
 import ru.practicum.shareit.item.dto.ItemForBookingDto;
-import ru.practicum.shareit.item.model.Item;
-import ru.practicum.shareit.user.User;
 import ru.practicum.shareit.user.dto.UserForBookingDto;
 
 import java.time.LocalDateTime;
@@ -17,14 +14,14 @@ import java.time.LocalDateTime;
 @Getter
 @Setter
 public class BookingDto {
-    private Integer id;
+    private Long id;
     private LocalDateTime start;
     private LocalDateTime end;
     private ItemForBookingDto item;
     private UserForBookingDto booker;
     private BookingStatus status;
 
-    public BookingDto(Integer id, LocalDateTime start, LocalDateTime end, BookingStatus status) {
+    public BookingDto(Long id, LocalDateTime start, LocalDateTime end, BookingStatus status) {
         this.id = id;
         this.start = start;
         this.end = end;
